BE/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
faiss-cpu==1.7.4
openai>=1.6.1
sentence-transformers>=2.5.0
numpy==1.24.3
pandas==2.0.3
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
aiofiles==23.2.1
python-magic==0.4.27
scikit-learn==1.3.2
requests==2.31.0
aiohttp==3.9.0
